+++ modP2pImpl/module-info.java
@@ -2,5 +2,5 @@ module aion.p2p.impl {
 	requires aion.p2p;
 	requires miniupnpc.linux;
 	requires java.xml.bind;
-	exports org.aion.p2p.impl;
+	exports org.aion.p2p.impl2;
 }
