+++ modP2pImpl/src/org/aion/p2p/impl/zero/msg/ReqActiveNodes.java
@@ -28,7 +28,7 @@ package org.aion.p2p.impl.zero.msg;
 import org.aion.p2p.Ctrl;
 import org.aion.p2p.Msg;
 import org.aion.p2p.Ver;
-import org.aion.p2p.impl.Act;
+import org.aion.p2p.impl.comm.Act;
 
 /**
  *
