+++ modP2pImpl/src/org/aion/p2p/impl/zero/msg/ResActiveNodes.java
@@ -33,8 +33,8 @@ import org.aion.p2p.Ctrl;
 import org.aion.p2p.INode;
 import org.aion.p2p.Msg;
 import org.aion.p2p.Ver;
-import org.aion.p2p.impl.Act;
-import org.aion.p2p.impl.Node;
+import org.aion.p2p.impl.comm.Act;
+import org.aion.p2p.impl.comm.Node;
 
 /**
  *
