+++ modP2pImpl/src/org/aion/p2p/impl2/selector/IOEventBus.java
@@ -1,4 +1,4 @@
-package org.aion.p2p.impl.selector;
+package org.aion.p2p.impl2.selector;
 
 import java.nio.channels.SelectableChannel;
 import java.util.Arrays;
